% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addInfos}
\alias{addInfos}
\title{adds two columns to a dataset}
\usage{
addInfos(dataset, productInfos)
}
\description{
This function adds row numbers to a dataset which equates an ID. Further a vector containing products' bulksize is added.
}
