% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSupplierDistance.R
\name{calculateSupplierDistance}
\alias{calculateSupplierDistance}
\title{calculate the Distance from the origin of a product to its supplier.}
\usage{
calculateSupplierDistance(originTable, producersTable)
}
\description{
from the data originTable which is the productOrigin and producersTable which is the producerAdress Table, the the distance form each products origin to its supplier is calculated.
}
