% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTotalDistances.R
\name{calculateTotalDistances}
\alias{calculateTotalDistances}
\title{calculate the total distances of the products.}
\usage{
calculateTotalDistances(origin = origin, producers = producerAdress,
  productInfo)
}
\arguments{
\item{origin}{the table productOrigin with the information ...}

\item{producers}{the table producerAdress}

\item{productInfo}{the table productInfo}
}
\description{
from the data productOrigin, producersAdress and productInfo a table will be
generated that contains the total distances for every product and its producer.
}
