% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDifference.R
\name{checkDifference}
\alias{checkDifference}
\title{check difference in data sets}
\usage{
checkDifference(dataset, productInfo)
}
\description{
This function checks if all the products of the given dataset are already known.
}
