% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentStorage.R
\name{currentStorage}
\alias{currentStorage}
\title{filters the current food storage}
\usage{
currentStorage(dataset, group)
}
\arguments{
\item{dataset}{kornumsatz as a result from \code{startupSettings()}}

\item{group}{optional character vector which contians product names}
}
\value{
A dataframe will be returned.
}
\description{
This function filters the current food storage. It looks for 
the last storage's change for each product and filters them. Optionally
one can give a \code{group} as input, which is a character strings 
containing product names. If this is given, the function will filter
the dataset by these product names.
}
