% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDataset.R
\name{editDataset}
\alias{editDataset}
\title{adds two columns to a dataset}
\usage{
editDataset(dataset, productInfo)
}
\description{
This function adds row numbers to a dataset which equates an ID. Further a vector containing products' bulksize is added.
}
