% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalise.R
\name{equalise}
\alias{equalise}
\title{generates an edited dataset}
\usage{
equalise(dataset, productInfo, reduce = T, pathTOmydb)
}
\description{
If \code{checkDifference} doesn't recognize any differences (anymore),
this function will generate an edited dataset, usually kornumsatz.
}
