% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyDate.R
\name{identifyDate}
\alias{identifyDate}
\title{identifies date}
\usage{
identifyDate(filename, type = "Date")
}
\arguments{
\item{filename}{character string of the filename which contains a date}

\item{type}{character string which defines the output, see details; default
is "Date"}
}
\value{
Depending on \code{type}, a \code{Date} or \code{character} will be 
returned
}
\description{
This function identifies the date of the given backup based on 
the file name. It separates it by underscores, dots and minus, substracts the 
essential numbers and converts it to a date. Optional one can choose the 
type and language of the printed date.
}
\details{
There are four defined output types defined at this moment. 
\enumerate{
  \item \code{"Date"} is from class \code{Date}
  \item \code{"numerical_EN"} looks like \code{"01 May 2018"}
  \item \code{"character_EN"} looks like \code{"1st of May 2018"}
  \item \code{"character_GER"} looks like \code{"1. Mai 2018"}
}
}
\examples{
identifyDate("Knk_2018_05_02--09:53.BAK", type = "character_GER")
}
