% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDatatable.R
\name{prepareDatatable}
\alias{prepareDatatable}
\title{prepare a datatable for shiny}
\usage{
prepareDatatable(kornumsatz, filter = "nothing", test = FALSE)
}
\arguments{
\item{kornumsatz}{kornumsatz data frame which results from 
\code{startupSettings()}}

\item{filter}{character string for filtering the dataset; default is
"nothing"; look at details}
}
\value{
A colored datatable is returned.
}
\description{
This function prepares a given dataset for visualization in
the foodstorage shiny app. It colors storage values below 20% in orange,
below 10% in red and more than 20% are colored in blue.
}
\details{
Filter can be: "nothing", which means no filter is applied.
"available" means only available products are shown. "empty" shows prducts
which are empty at this moment.
}
