% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosIs.R
\name{prognosIs}
\alias{prognosIs}
\title{function for calculating a regression}
\usage{
prognosIs(product, from = "", to = "", graphics = TRUE, type = "p",
  col_points = "grey", pch = 16, main_header = "",
  ylab = "Warenbestand in Kilo", xlab = "", las = 1, col_reg = "grey",
  col_conv = "lightgrey", col_20 = "red", col_past = "black",
  smoother = "loess", span = 0.1, degree = 1, lwd = c(2, 2, 1),
  lty = c(1, 1, 2), nec.dates = 10, more.than = 15, test = FALSE)
}
\arguments{
\item{product}{Need to be a character string}

\item{from}{As default plot will be generated for the last year}

\item{to}{As default end of time serie is the date of last backup import}

\item{graphics}{logical. If it's true, a plot will be the result. Otherwise a dataframe containing dates}

\item{type}{Choose type how points shall be drawn. Visible as default.}

\item{col_points}{If points are drawn, choose color here. Default is grey.}

\item{pch}{same as usual... As default filled points are drawn.}

\item{main_header}{Not necessary, but it's recommended for German language}

\item{ylab}{As default 'food storage in kilo' in German}

\item{xlab}{Can be choosen, but not necessary. It's generated automatically}

\item{las}{Of course numbers of y-axis are rotated about 90 degrees}

\item{col_reg}{Optional: Choose regression line's color}

\item{col_conv}{Optional: Choose color of lines of confidential interval}

\item{col_20}{Optional: Choose 'run-out' line's color}

\item{col_past}{Optional: Choose the color for loess line}

\item{smoother}{At this moment there is implemented only loess}

\item{span}{Attributes for loess. As default it's set as 0.1}

\item{degree}{As default it's 1}

\item{lwd}{Three numbers in one vector are necessary. Order is loess-regression-conf.interval}

\item{lty}{Same as lwd}

\item{nec.dates}{How many days do you want at least for calculating a regression?}

\item{more.than}{Important for correct giving data. Default should be ok}

\item{test}{logical. If TRUE you can test fun_reg by test_function {foodstorage}}
}
\value{
If graphics is true (default), the result will be a plot. X-axis consists of a time serie, y-axis consists of the daily food storage. If graphics is false, a list is spewed containing two dates: The second one is the date, when your storage will run out of the giving product. And the first one is the date four weeks before end.
}
\description{
function for calculating a regression
}
