% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStorageApp.R
\name{runStorageApp}
\alias{runStorageApp}
\title{function for loading the storage-app}
\usage{
runStorageApp()
}
\value{
This function is the only exported function of the foodstorage package. The package was built for analyzing a Food Co-ops' consumption. The results are represented in the shiny 'storage-app' which you can run by this function!
}
\description{
function for loading the storage-app
}
