% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startupSettings.R
\name{startupSettings}
\alias{startupSettings}
\title{changes the input dataset}
\usage{
startupSettings(dataset, productInfo)
}
\description{
This function edits the input dataset by adding more information about specific
products like bulksize. The information comes from a second dataset, productInfo, which needs
to have a column containing all names of products that exist in the original dataset.
}
