% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoverPerYear.R
\name{turnoverPerYear}
\alias{turnoverPerYear}
\title{calculate the quantities of the products per year.}
\usage{
turnoverPerYear(kornumsatz, productInfo, group_var)
}
\description{
from the data kornumsatz, which has the columns "Tag", "Menge", "Produkt" and the data productInfo, with the columns "Produkte_App", "Produkte_Zusammenfassung", the kornumsatz (quantity of products) for every product and year.
}
\examples{
turnoverPerYear(kornumsatz, productInfo,  Produkte_Zusammenfassung)
}
